Add support for named postgres types.
Upon connection, we query the database for a list of all types it knows about.
We then create a hash mapping their OID->type name. This allows types to be registered
by name, which is stable across all installations, as opposed to registering by OID
which varies between databases even within the same cluster.

This adds a hash lookup indirection and an extra method call, which need profiling
to prove they don't add an undue cost. It also preserves the traditional PG_TYPE
registration method for backwards compatibility with existing plugins and modules.